Cleaned code, less repeating. Will change prop drill to global state later
