+++ frontend/src/types/index.ts
@@ -7,3 +7,12 @@ export interface Vinyl {
 	coverImg: string;
 	owned: boolean;
 }
+
+export interface Props {
+	vinyls: Vinyl[];
+	search: string;
+	genre: string;
+	setVinyls: React.Dispatch<React.SetStateAction<Vinyl[]>>;
+	setSearch: React.Dispatch<React.SetStateAction<string>>;
+	setGenre: React.Dispatch<React.SetStateAction<string>>;
+}
